refactor: reorganize project structure
- Adopt cmd/internal layout
- Move root files into appropriate packages
- Consolidate database logic
- Update import paths
